* Allow key-like items from the consumer page to be recognized as well. * Make the key arrays take the full range of possible index values (32-bits). * Forward all unmapped keys with their HID usage as keycode. * Remove some superflous debug output.
With this many of the extra keys on keyboards should be forwarded as unmapped
keys. An application could now watch for them (B_UNMAPPED_KEY{UP|DOWN} messages)
and interpret them based on their keycode which matches the HID usage.


git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@31838 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/drivers/input/usb_hid/KeyboardDevice.h
@@ -45,8 +45,8 @@ private:
 		uint32					fModifierCount;
 
 		uint8					fLastModifiers;
-		uint8 *					fCurrentKeys;
-		uint8 *					fLastKeys;
+		uint32 *				fCurrentKeys;
+		uint32 *				fLastKeys;
 };
 
 #endif // USB_KEYBOARD_DEVICE_H
